ProjectSetup

+++ lab9/DentalSurgeries/src/main/java/org/cs489/dentalsurgeries/repository/SurgeryRepository.java
@@ -3,5 +3,5 @@ package org.cs489.dentalsurgeries.repository;
 import org.cs489.dentalsurgeries.model.Surgery;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface SurgeryRepository extends JpaRepository<Surgery, Integer> {
+public interface SurgeryRepository extends JpaRepository<Surgery, Long> {
 }
